Original change: https://android-review.googlesource.com/c/device/generic/vulkan-cereal/+/1836676

+++ stream-servers/virtio-gpu-gfxstream-renderer.cpp
@@ -927,7 +927,7 @@ public:
         return 0;
     }
 
-    int contextCreateFence(uint64_t fence_id, uint32_t ctx_id, uint32_t fence_ctx_idx) {
+    int contextCreateFence(uint64_t fence_id, uint32_t ctx_id, uint8_t fence_ctx_idx) {
         AutoLock lock(mLock);
         VGPLOG("fenceid: %llu cmdtype: %u fence_ctx_idx: %u", (unsigned long long)fence_id, ctx_id, fence_ctx_idx);
         if (mVirtioGpuTimelines) {
@@ -1807,7 +1807,7 @@ VG_EXPORT int stream_renderer_resource_unmap(uint32_t res_handle) {
 }
 
 VG_EXPORT int stream_renderer_context_create_fence(
-    uint64_t fence_id, uint32_t ctx_id, uint32_t fence_ctx_idx) {
+    uint64_t fence_id, uint32_t ctx_id, uint8_t fence_ctx_idx) {
     sRenderer()->contextCreateFence(fence_id, ctx_id, fence_ctx_idx);
     return 0;
 }
